introduce refreshing; frontend: add update() functions
Add a refreshing functionality.

In the frontend, there is now a update() function in most of the classes
that replaces the various sizechange/timechange/redraw ... functions.
Other classes shall now call only call update with an appropriate
`changes` parameter and the classes choose which actions to take
themselves. (This transition is not 100% complete yet at this time.)